+++ tests/test_runs.py
@@ -7,6 +7,7 @@ import sys
 import subprocess
 from django.core.management import call_command
 from django.test import TestCase
+import pexpect
 
 
 class TestDeploy(TestCase):
@@ -56,3 +57,37 @@ def test_subprocess_opt_error():
         "CommandError: Failed to convert collectstatic options to CLI format: {'not-an-option': False}"
         in result.stderr
     )
+
+
+def test_option_toggle():
+    child = pexpect.spawn(
+        " ".join(
+            [
+                sys.executable,
+                "./manage.py",
+                "routine",
+                "--settings",
+                "tests.settings_option_toggle",
+                "option-on",
+                "--subprocess",
+            ]
+        )
+    )
+    child.expect("static files copied.")
+
+    child = pexpect.spawn(
+        " ".join(
+            [
+                sys.executable,
+                "./manage.py",
+                "routine",
+                "--settings",
+                "tests.settings_option_toggle",
+                "option-off",
+                "--subprocess",
+            ]
+        )
+    )
+    child.expect("Type 'yes' to continue, or 'no' to cancel:")
+    child.sendline("yes")
+    child.expect("static files copied.")