+++ doc/source/changelog.rst
@@ -8,6 +8,7 @@ v1.3.0 (2024-02-18)
 
 * Implemented `Remove support for python 3.8 <https://github.com/bckohan/django-render-static/issues/30>`_
 * Implemented `Upgrade to django-typer 3.x <https://github.com/bckohan/django-render-static/issues/29>`_
+* Implemented `stdout/stderr streaming from subprocesses <https://github.com/bckohan/django-render-static/issues/15>`_
 
 v1.2.1 (2024-08-26)
 ===================